Take out certs and errors
