+++ scripts/deploy/src/error.js
@@ -1,0 +1,9 @@
+
+class UserRuntimeError extends Error {
+    constructor(message) {
+        super(message);
+        this.name = "UserRuntimeError";
+    }
+}
+
+export default UserRuntimeError;
